API_Commandes/requirements.txt
fastapi>=0.95.0 # Pour le framework web
uvicorn>=0.21.1 # Pour le lancement de l'API
sqlalchemy #Pour la partie BDD
pydantic>=1.10.7 # Pour la validation des données
pytest>=7.3.1 # Pour les tests
httpx>=0.24.0  # Pour FastAPI TestClient
pylint # Pour l'analyse statique du code
python-dotenv # Pour la gestion des variables d'environnement
requests # Pour les requêtes HTTP
pika>=1.3.1 # Pour la communication avec RabbitMQ
pydantic[email] # Pour la validation des données avec des emails
pyyaml # Pour la gestion des fichiers de configuration YAML
prometheus_client # Pour la collecte de métriques
python-jose[cryptography] # Pour la gestion des JWT
passlib[bcrypt] # Pour le hachage des mots de passe
psycopg2-binary # Pour la connexion à PostgreSQL